+++ sample/live/liveMedia/MPEG4VideoStreamFramer.cpp
@@ -651,8 +651,9 @@ unsigned MPEG4VideoStreamParser::parseVideoObjectPlane() {
       }
       setParseState(PARSING_VIDEO_OBJECT_PLANE);
     } else {
+      /* add (unsigned long) cast for -Wint-to-pointer-cast warning of compiler */
       usingSource()->envir() << "MPEG4VideoStreamParser::parseVideoObjectPlane(): Saw unexpected code "
-			     << (void*)next4Bytes << "\n";
+			     << (void*)((unsigned long)next4Bytes) << "\n";
       setParseState(PARSING_VIDEO_OBJECT_PLANE); // the safest way to recover...
     }
     break;
